refactor: divide `polygon.py` into `point.py` and `box.py`
