+++ src/shapepy/bool2d/primitive.py
@@ -17,7 +17,7 @@ import numpy as np
 from shapepy.bool2d.shape import EmptyShape, SimpleShape, WholeShape
 from shapepy.geometry.curve import PlanarCurve
 from shapepy.geometry.jordancurve import JordanCurve
-from shapepy.geometry.polygon import Point2D
+from shapepy.geometry.point import Point2D
 
 
 class Primitive:
