+++ src/shapepy/geometry/curve.py
@@ -20,7 +20,8 @@ from typing import Any, Optional, Tuple, Union
 import numpy as np
 import pynurbs
 
-from shapepy.geometry.polygon import Box, Point2D
+from shapepy.geometry.box import Box
+from shapepy.geometry.point import Point2D
 
 
 class Math:
