+++ src/shapepy/geometry/jordancurve.py
@@ -11,8 +11,9 @@ from typing import Optional, Tuple, Union
 
 import numpy as np
 
+from shapepy.geometry.box import Box
 from shapepy.geometry.curve import IntegratePlanar, PlanarCurve
-from shapepy.geometry.polygon import Box, Point2D
+from shapepy.geometry.point import Point2D
 
 
 class IntegrateJordan:
